move more logic to utils
